backend/requirements.txt
fastapi>=0.115.6,<1.0
uvicorn[standard]>=0.32.1,<1.0
pydantic>=2.10.3,<3.0
pydantic-settings>=2.7.0,<3.0
httpx>=0.28.1,<1.0
orjson>=3.10.12,<4.0
python-multipart>=0.0.20,<1.0
pypdf>=5.1.0,<6.0
python-docx>=1.1.2,<2.0
sqlmodel>=0.0.22,<1.0
sqlalchemy[asyncio]>=2.0.36,<3.0
asyncpg>=0.30.0,<1.0
redis>=5.2.1,<6.0
greenlet>=3.0.3,<3.2
